Move register_in_track function to the util
